requirements.txt
Django==3.1.4
djangorestframework==3.12.2
celery==5.0.4
redis==3.5.3
